requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pymongo>=4.9
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
groq==0.28.0
pydantic==2.11.7
python-multipart==0.0.6
boto3==1.28.38
gTTS==2.3.2

# Additional dependencies for enhanced functionality
email-validator==2.1.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Trusted hosts for installation
--trusted-host pypi.org
--trusted-host pypi.python.org
--trusted-host files.pythonhosted.org